Snap 2562 (#436)
* Validate-disk-store tool was broken for co-location check.
Reason - We create a "SNAPPY_INTERNAL_DELTA" for row buffer pr region and the column store disk store in the default disk store. This tool was modified to check whether all the related buckets are there or not and the assumption was that all the colocated PRs will use the same DISK STORE for persistence.
This assumption is not right and in case of column tables we have actually separated the row buffer pr disk store and the column buffer pr diskstore.

So right now the tool has been modified to also list all the disk regions, so that at least manually it can be verified.