Snap 3193 (#523)
* Fix for bug SNAP-3193. If the precision type of Float is not specified, then the default is taken to be max precsion for Real data type instead of double

* Supressing throwing of exception if the value in Real data type is outside its precision & getting rounded to 0

The issue is that if a Float column type Column is present in DDL, then if the table is created as column table, the data type of column is correctly present as float. But if a row table is created, then the data type is double. This is because for row table in gfxd, the float is mapped to REAL type. And a Real Type without precision is assumed to be a Double Precision in sqlGrammar.jj.
The first change is that in sqlGrammar.jj where if the precision type with Real is not specified it is taken as max precision allowed for REAL rather than Double's precision, if the system is a snappy store.
The second change is in SQLReal, where if the value passed is a double, and rounding off occurs, then exception is thrown, for row table. But for column table , the double is rounded to 0 to fit in the Float data type. So the change is that in case of snappy store , do not thrown an exception if rounding occurs.
Added bug test in
cluster org.apache.spark.sql.store.BugTest